tyre wear and weather conditions are ok.
