+++ src/PitWallDataGatheringApi/PitWallDataGatheringApi.Integration.Tests/PitWallDataGatheringApi.Integration.Tests.v3.ncrunchproject
@@ -4,9 +4,6 @@
       <NamedTestSelector>
         <TestName>PitWallDataGatheringApi.Integration.Tests.InstantMetricReadTest.GIVEN_metric_THEN_read_from_timeSerie</TestName>
       </NamedTestSelector>
-      <FixtureTestSelector>
-        <FixtureName>PitWallDataGatheringApi.Integration.Tests.InstantMetricReadTest</FixtureName>
-      </FixtureTestSelector>
     </IgnoredTests>
   </Settings>
 </ProjectConfiguration>
